Non-destructive subject loading
